backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
pillow==11.0.0
# opencv-python-headless==4.10.0.84  # Commented out for faster install
openai==1.54.0
httpx<0.28.0  # Pin to <0.28 for openai compatibility (0.28+ removed 'proxies' param)
anthropic==0.72.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
aiofiles==24.1.0
redis==5.2.0
celery==5.4.0

# Authentication dependencies
bcrypt==4.2.1
passlib==1.7.4
pyjwt==2.9.0
email-validator==2.2.0
